@@ -10,8 +10,8 @@
 	},
 	"dependencies": {
 		"core-js": "^3.6.4",
-		"vue": "^2.6.11",
-		"vue-router": "^3.1.6"
+		"vue": "^3.0.0-beta.1",
+		"vue-router": "^4.0.0-alpha.5"
 	},
 	"devDependencies": {
 		"@types/jest": "^24.0.19",
@@ -23,16 +23,17 @@
 		"@vue/cli-plugin-typescript": "~4.3.0",
 		"@vue/cli-plugin-unit-jest": "~4.3.0",
 		"@vue/cli-service": "~4.3.0",
+		"@vue/compiler-sfc": "^3.0.0-beta.1",
 		"@vue/eslint-config-prettier": "^6.0.0",
 		"@vue/eslint-config-typescript": "^5.0.2",
-		"@vue/test-utils": "1.0.0-beta.31",
+		"@vue/test-utils": "^2.0.0-alpha.1",
 		"eslint": "^6.7.2",
 		"eslint-plugin-prettier": "^3.1.1",
-		"eslint-plugin-vue": "^6.2.2",
+		"eslint-plugin-vue": "^7.0.0-alpha.0",
 		"prettier": "^1.19.1",
 		"sass": "^1.26.3",
 		"sass-loader": "^8.0.2",
 		"typescript": "~3.8.3",
-		"vue-template-compiler": "^2.6.11"
+		"vue-cli-plugin-vue-next": "~0.1.2"
 	}
 }
