@@ -1,10 +1,9 @@
-import Vue from "vue";
-import App from "./App.vue";
-import router from "./router";
+import { createApp } from 'vue'
+import App from './App.vue'
+import router from './router'
+import './theme/src/index.scss'
+import './views/demo.css'
 
-Vue.config.productionTip = false;
-
-new Vue({
-	router,
-	render: h => h(App)
-}).$mount("#app");
+createApp(App as any)
+	.use(router)
+	.mount('#app')
