@@ -4,28 +4,28 @@ module.exports = {
 		node: true
 	},
 	extends: [
-		"plugin:vue/essential",
-		"eslint:recommended",
-		"@vue/typescript/recommended",
-		"@vue/prettier",
-		"@vue/prettier/@typescript-eslint"
+		'plugin:vue/vue3-essential',
+		'eslint:recommended',
+		'@vue/typescript/recommended',
+		'@vue/prettier',
+		'@vue/prettier/@typescript-eslint'
 	],
 	parserOptions: {
 		ecmaVersion: 2020
 	},
 	rules: {
-		"no-console": process.env.NODE_ENV === "production" ? "warn" : "off",
-		"no-debugger": process.env.NODE_ENV === "production" ? "warn" : "off"
+		'no-console': process.env.NODE_ENV === 'production' ? 'warn' : 'off',
+		'no-debugger': process.env.NODE_ENV === 'production' ? 'warn' : 'off',
+		quotes: ['error', 'single'],
+		semi: ['error', 'never'],
+		'@typescript-eslint/no-explicit-any': 'off'
 	},
 	overrides: [
 		{
-			files: [
-				"**/__tests__/*.{j,t}s?(x)",
-				"**/tests/unit/**/*.spec.{j,t}s?(x)"
-			],
+			files: ['**/__tests__/*.{j,t}s?(x)', '**/tests/unit/**/*.spec.{j,t}s?(x)'],
 			env: {
 				jest: true
 			}
 		}
 	]
-};
+}
