@@ -1,28 +1,35 @@
-import Vue from "vue";
-import VueRouter, { RouteConfig } from "vue-router";
-import Home from "../views/Home.vue";
+import { createRouter, createWebHashHistory } from 'vue-router'
+import Home from '../views/Home.vue'
 
-Vue.use(VueRouter);
-
-const routes: Array<RouteConfig> = [
+const routes: Array<any> = [
 	{
-		path: "/",
-		name: "Home",
+		path: '/',
+		name: 'Home',
 		component: Home
 	},
 	{
-		path: "/about",
-		name: "About",
+		path: '/about',
+		name: 'About',
 		// route level code-splitting
 		// this generates a separate chunk (about.[hash].js) for this route
 		// which is lazy-loaded when the route is visited.
-		component: () =>
-			import(/* webpackChunkName: "about" */ "../views/About.vue")
+		component: () => import(/* webpackChunkName: "about" */ '../views/About.vue')
+	},
+	{
+		path: '/button',
+		name: 'Button',
+		component: () => import(/* webpackChunkName: "button" */ '../views/Button.vue')
+	},
+	{
+		path: '/radio',
+		name: 'Radio',
+		component: () => import(/* webpackChunkName: "radio" */ '../views/Radio.vue')
 	}
-];
+]
 
-const router = new VueRouter({
+const router = createRouter({
+	history: createWebHashHistory(),
 	routes
-});
+})
 
-export default router;
+export default router
