@@ -1,4 +1,9 @@
-declare module "*.vue" {
-	import Vue from "vue";
-	export default Vue;
+declare module '*.vue' {
+	// import Vue from 'vue'
+	import * as Vue from 'vue'
+	export default Vue
+}
+
+declare interface Data {
+	[key: string]: any
 }
